[REFACTOR/#123] 로그인 모듈 설정
