+++ feature/auth/src/main/java/co/orange/auth/phone/TermBottomSheet.kt
@@ -1,17 +1,15 @@
-package co.orange.presentation.auth.phone
+package co.orange.auth.phone
 
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
+import co.orange.auth.databinding.BottomSheetTermBinding
 import co.orange.core.R
 import co.orange.core.base.BaseBottomSheet
 import co.orange.core.extension.setOnSingleClickListener
-import co.orange.presentation.setting.SettingActivity.Companion.WEB_TERM_PRIVATE
-import co.orange.presentation.setting.SettingActivity.Companion.WEB_TERM_SERVICE
-import kr.genti.presentation.databinding.BottomSheetTermBinding
-import kr.genti.presentation.R as featureR
+import co.orange.auth.R as featureR
 
 class TermBottomSheet :
     BaseBottomSheet<BottomSheetTermBinding>(featureR.layout.bottom_sheet_term) {
@@ -53,5 +51,12 @@ class TermBottomSheet :
             viewModel.clickSubmitBtn(true)
             dismiss()
         }
+    }
+
+    companion object {
+        const val WEB_TERM_PRIVATE =
+            "https://brawny-guan-098.notion.site/5a8b57e78f594988aaab08b8160c3072?pvs=4"
+        const val WEB_TERM_SERVICE =
+            "https://brawny-guan-098.notion.site/faa1517ffed44f6a88021a41407ed736?pvs=4"
     }
 }
