+++ feature/auth/src/main/res/layout/activity_phone.xml
